Check cart is editable before creating/updating/deleting a cart item

+++ modules/ocApiCartItems/actions/actions.class.php
@@ -61,12 +61,12 @@ class ocApiCartItemsActions extends apiActions
      */
     public function create(sfWebRequest $request)
     {
+        $cart_id = $request->getParameter('id', 0);
+
         /* @var $cartsService ApiCartsService */
         $cartsService = $this->getService('carts_service');
-        $cart_id = $request->getParameter('id', 0);
-        $cart = $cartsService->findOneById($cart_id);
-        if ( 0 == count($cart) ) {
-            return $this->createBadRequestResponse(['error' => 'Cart not found with id=' . $cart_id]);
+        if (!$cartsService->isCartEditable($cart_id)) {
+            return $this->createBadRequestResponse(['error' => "Cart not found or not editable (id=$cart_id)"]);
         }
 
         /* @var $cartItemsService ApiCartItemsService */
@@ -94,9 +94,15 @@ class ocApiCartItemsActions extends apiActions
         $cart_id = $request->getParameter('id', 0);
         $item_id = $request->getParameter('item_id', 0);
 
-        /* @var $cartService ApiCartsService */
-        $cartService = $this->getService('cartitems_service');
-        $isSuccess = $cartService->updateCartItem($cart_id, $item_id, $request->getPostParameters());
+        /* @var $cartsService ApiCartsService */
+        $cartsService = $this->getService('carts_service');
+        if (!$cartsService->isCartEditable($cart_id)) {
+            return $this->createBadRequestResponse(['error' => "Cart not found or not editable (id=$cart_id)"]);
+        }
+
+        /* @var $cartItemsService ApiCartItemsService */
+        $cartItemsService = $this->getService('cartitems_service');
+        $isSuccess = $cartItemsService->updateCartItem($cart_id, $item_id, $request->getPostParameters());
 
         if (!$isSuccess) {
             $status = ApiHttpStatus::BAD_REQUEST;
@@ -123,6 +129,12 @@ class ocApiCartItemsActions extends apiActions
         $cart_id = $request->getParameter('id', 0);
         $item_id = $request->getParameter('item_id', 0);
 
+        /* @var $cartsService ApiCartsService */
+        $cartsService = $this->getService('carts_service');
+        if (!$cartsService->isCartEditable($cart_id)) {
+            return $this->createBadRequestResponse(['error' => "Cart not found or not editable (id=$cart_id)"]);
+        }
+
         /* @var $cartItemsService ApiCartItemsService */
         $cartItemsService = $this->getService('cartitems_service');
         $isSuccess = $cartItemsService->deleteCartItem($cart_id, $item_id);
@@ -136,4 +148,6 @@ class ocApiCartItemsActions extends apiActions
                 'message' => $message
                 ], $status);
     }
+
+
 }
